requirements.txt
APScheduler==2.1.2
certifi==2021.5.30
chardet==4.0.0
idna==2.10
python-dotenv==0.17.1
requests==2.25.1
